requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sentence-transformers
faiss-cpu
openai
numpy
